+++ source/Energinet.DataHub.MarketParticipant.Application/Validation/UpdateActorCommandRuleSet.cs
@@ -12,8 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.MarketParticipant.Application.Commands;
-using Energinet.DataHub.MarketParticipant.Application.Validation.Rules;
+using Energinet.DataHub.MarketParticipant.Application.Commands.Actor;
 using Energinet.DataHub.MarketParticipant.Domain.Model;
 using FluentValidation;
 
@@ -24,27 +23,34 @@ namespace Energinet.DataHub.MarketParticipant.Application.Validation
         public UpdateActorCommandRuleSet()
         {
             RuleFor(command => command.OrganizationId)
-                .NotEmpty()
-                .SetValidator(new GuidValidationRule<UpdateActorCommand>());
+                .NotEmpty();
 
             RuleFor(command => command.ActorId)
-                .NotEmpty()
-                .SetValidator(new GuidValidationRule<UpdateActorCommand>());
+                .NotEmpty();
 
-            RuleFor(actor => actor.MarketRoles)
+            RuleFor(actor => actor.ChangeActor)
                 .NotNull()
-                .ChildRules(rolesValidator =>
+                .ChildRules(changeActorValidator =>
                 {
-                    rolesValidator
-                        .RuleForEach(x => x)
+                    changeActorValidator
+                        .RuleFor(x => x.Status)
+                        .NotEmpty()
+                        .IsEnumName(typeof(ActorStatus));
+
+                    changeActorValidator
+                        .RuleFor(x => x.MarketRoles)
                         .NotNull()
-                        .ChildRules(roleValidator =>
-                        {
-                            roleValidator
-                                .RuleFor(x => x.Function)
-                                .NotEmpty()
-                                .IsEnumName(typeof(EicFunction), false);
-                        });
+                        .ChildRules(rolesValidator =>
+                            rolesValidator
+                                .RuleForEach(x => x)
+                                .NotNull()
+                                .ChildRules(roleValidator =>
+                                {
+                                    roleValidator
+                                        .RuleFor(x => x.Function)
+                                        .NotEmpty()
+                                        .IsEnumName(typeof(EicFunction), false);
+                                }));
                 });
         }
     }