+++ source/Energinet.DataHub.MarketParticipant.Tests/Validation/CreateActorCommandRuleSetTests.cs
@@ -15,7 +15,7 @@
 using System;
 using System.Linq;
 using System.Threading.Tasks;
-using Energinet.DataHub.MarketParticipant.Application.Commands;
+using Energinet.DataHub.MarketParticipant.Application.Commands.Actor;
 using Energinet.DataHub.MarketParticipant.Application.Validation;
 using Xunit;
 using Xunit.Categories;
@@ -35,7 +35,7 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
             const string propertyName = nameof(CreateActorCommand.Actor);
 
             var target = new CreateActorCommandRuleSet();
-            var command = new CreateActorCommand(ValidId, null!);
+            var command = new CreateActorCommand(Guid.Parse(ValidId), null!);
 
             // Act
             var result = await target.ValidateAsync(command).ConfigureAwait(false);
@@ -46,20 +46,16 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
         }
 
         [Theory]
-        [InlineData("", false)]
-        [InlineData(null, false)]
-        [InlineData("  ", false)]
         [InlineData("8F9B8218-BAE6-412B-B91B-0C78A55FF128", true)]
-        [InlineData("8F9B8218-BAE6-412B-B91B-0C78A55FF1XX", false)]
         public async Task Validate_OrganizationId_ValidatesProperty(string value, bool isValid)
         {
             // Arrange
             const string propertyName = nameof(CreateActorCommand.OrganizationId);
 
-            var actorDto = new ChangeActorDto(new GlobalLocationNumberDto(ValidGln), Array.Empty<MarketRoleDto>());
+            var actorDto = new CreateActorDto(new GlobalLocationNumberDto(ValidGln), Array.Empty<MarketRoleDto>());
 
             var target = new CreateActorCommandRuleSet();
-            var command = new CreateActorCommand(value, actorDto);
+            var command = new CreateActorCommand(Guid.Parse(value), actorDto);
 
             // Act
             var result = await target.ValidateAsync(command).ConfigureAwait(false);
@@ -85,12 +81,12 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
         public async Task Validate_ActorGln_ValidatesProperty(string value, bool isValid)
         {
             // Arrange
-            var propertyName = $"{nameof(CreateActorCommand.Actor)}.{nameof(ChangeActorDto.Gln)}";
+            var propertyName = $"{nameof(CreateActorCommand.Actor)}.{nameof(CreateActorDto.Gln)}";
 
-            var actorDto = new ChangeActorDto(new GlobalLocationNumberDto(value), Array.Empty<MarketRoleDto>());
+            var actorDto = new CreateActorDto(new GlobalLocationNumberDto(value), Array.Empty<MarketRoleDto>());
 
             var target = new CreateActorCommandRuleSet();
-            var command = new CreateActorCommand(ValidId, actorDto);
+            var command = new CreateActorCommand(Guid.Parse(ValidId), actorDto);
 
             // Act
             var result = await target.ValidateAsync(command).ConfigureAwait(false);
@@ -114,10 +110,10 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
             // Arrange
             var propertyName = $"{nameof(CreateActorCommand.Actor)}.{nameof(ChangeActorDto.MarketRoles)}";
 
-            var organizationRoleDto = new ChangeActorDto(new GlobalLocationNumberDto(ValidGln), null!);
+            var organizationRoleDto = new CreateActorDto(new GlobalLocationNumberDto(ValidGln), null!);
 
             var target = new CreateActorCommandRuleSet();
-            var command = new CreateActorCommand(ValidId, organizationRoleDto);
+            var command = new CreateActorCommand(Guid.Parse(ValidId), organizationRoleDto);
 
             // Act
             var result = await target.ValidateAsync(command).ConfigureAwait(false);
@@ -133,10 +129,10 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
             // Arrange
             var propertyName = $"{nameof(CreateActorCommand.Actor)}.{nameof(ChangeActorDto.MarketRoles)}[0]";
 
-            var organizationRoleDto = new ChangeActorDto(new GlobalLocationNumberDto(ValidGln), new MarketRoleDto[] { null! });
+            var organizationRoleDto = new CreateActorDto(new GlobalLocationNumberDto(ValidGln), new MarketRoleDto[] { null! });
 
             var target = new CreateActorCommandRuleSet();
-            var command = new CreateActorCommand(ValidId, organizationRoleDto);
+            var command = new CreateActorCommand(Guid.Parse(ValidId), organizationRoleDto);
 
             // Act
             var result = await target.ValidateAsync(command).ConfigureAwait(false);
@@ -160,10 +156,10 @@ namespace Energinet.DataHub.MarketParticipant.Tests.Validation
             // Arrange
             var propertyName = $"{nameof(CreateActorCommand.Actor)}.{nameof(ChangeActorDto.MarketRoles)}[0].{nameof(MarketRoleDto.Function)}";
 
-            var organizationRoleDto = new ChangeActorDto(new GlobalLocationNumberDto(ValidGln), new[] { new MarketRoleDto(value) });
+            var organizationRoleDto = new CreateActorDto(new GlobalLocationNumberDto(ValidGln), new[] { new MarketRoleDto(value) });
 
             var target = new CreateActorCommandRuleSet();
-            var command = new CreateActorCommand(ValidId, organizationRoleDto);
+            var command = new CreateActorCommand(Guid.Parse(ValidId), organizationRoleDto);
 
             // Act
             var result = await target.ValidateAsync(command).ConfigureAwait(false);