+++ source/Energinet.DataHub.MarketParticipant.Client/documents/release-notes/release-notes.md
@@ -1,5 +1,28 @@
 # Energinet.DataHub.MarketParticipant.Client Release notes
 
-## Version 0.1.0
+## Version 0.2.1
 
 - Preparing package for release.
+
+## Version 0.3.1
+
+### Added methods for the following endpoints
+
+- For actors
+    - GetActorsAsync
+    - GetActorAsync
+    - CreateActorAsync
+    - UpdateActorAsync
+- For Organizations
+    - GetOrganizationsAsync
+    - GetOrganizationAsync
+    - CreateOrganizationAsync
+    - UpdateOrganizationAsync
+
+## Version 0.4.0
+
+- Global location number cannot be changed on an actor.
+
+## Version 0.4.1
+
+- Added CVR and Address to organization